requirements.txt
Flask==2.3.2
python-pptx==0.6.21
playwright==1.48.0
Pillow==11.0.0
